Enable/disable timer, small refactoring
